--- .vs/ClassBooking/v17/DocumentLayout.json
@@ -1,60 +1,0 @@
-{
-  "Version": 1,
-  "WorkspaceRootPath": "C:\\Users\\mire0\\Documents\\coding\\ClassBooking\\ClassBooking\\",
-  "Documents": [
-    {
-      "AbsoluteMoniker": "D:0:0:{013E1B04-E320-4AB6-97DA-D52238E83783}|ClassBooking\\ClassBooking.vcxproj|C:\\Users\\mire0\\Documents\\coding\\ClassBooking\\ClassBooking\\ClassBooking\\ClassBooking.cpp||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}",
-      "RelativeMoniker": "D:0:0:{013E1B04-E320-4AB6-97DA-D52238E83783}|ClassBooking\\ClassBooking.vcxproj|solutionrelative:ClassBooking\\ClassBooking.cpp||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}"
-    },
-    {
-      "AbsoluteMoniker": "D:0:0:{A2FE74E1-B743-11D0-AE1A-00A0C90FFFC3}|\u003CMiscFiles\u003E|C:\\Users\\mire0\\Documents\\coding\\ClassBooking\\ClassBooking\\ClassBooking.cpp||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}",
-      "RelativeMoniker": "D:0:0:{A2FE74E1-B743-11D0-AE1A-00A0C90FFFC3}|\u003CMiscFiles\u003E|solutionrelative:ClassBooking.cpp||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}"
-    }
-  ],
-  "DocumentGroupContainers": [
-    {
-      "Orientation": 0,
-      "VerticalTabListWidth": 256,
-      "DocumentGroups": [
-        {
-          "DockedWidth": 200,
-          "SelectedChildIndex": 2,
-          "Children": [
-            {
-              "$type": "Bookmark",
-              "Name": "ST:0:0:{3ae79031-e1bc-11d0-8f78-00a0c9110057}"
-            },
-            {
-              "$type": "Bookmark",
-              "Name": "ST:0:0:{cce594b6-0c39-4442-ba28-10c64ac7e89f}"
-            },
-            {
-              "$type": "Document",
-              "DocumentIndex": 0,
-              "Title": "ClassBooking.cpp",
-              "DocumentMoniker": "C:\\Users\\mire0\\Documents\\coding\\ClassBooking\\ClassBooking\\ClassBooking\\ClassBooking.cpp",
-              "RelativeDocumentMoniker": "ClassBooking\\ClassBooking.cpp",
-              "ToolTip": "C:\\Users\\mire0\\Documents\\coding\\ClassBooking\\ClassBooking\\ClassBooking\\ClassBooking.cpp",
-              "RelativeToolTip": "ClassBooking\\ClassBooking.cpp",
-              "ViewState": "AQIAAAAAAAAAAAAAAAAAAAQAAAAAAAAA",
-              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000677|",
-              "WhenOpened": "2025-04-14T06:16:23.635Z",
-              "EditorCaption": ""
-            },
-            {
-              "$type": "Document",
-              "DocumentIndex": 1,
-              "Title": "ClassBooking.cpp",
-              "DocumentMoniker": "C:\\Users\\mire0\\Documents\\coding\\ClassBooking\\ClassBooking\\ClassBooking.cpp",
-              "RelativeDocumentMoniker": "ClassBooking.cpp",
-              "ToolTip": "C:\\Users\\mire0\\Documents\\coding\\ClassBooking\\ClassBooking\\ClassBooking.cpp",
-              "RelativeToolTip": "ClassBooking.cpp",
-              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000677|",
-              "WhenOpened": "2025-04-14T05:59:01.086Z"
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}
