+++ ClassBooking.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.9.34714.143
+VisualStudioVersion = 17.11.35312.102
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ClassBooking", "ClassBooking\ClassBooking.vcxproj", "{013E1B04-E320-4AB6-97DA-D52238E83783}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ClassBooking", "ClassBooking.vcxproj", "{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,19 +13,19 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{013E1B04-E320-4AB6-97DA-D52238E83783}.Debug|x64.ActiveCfg = Debug|x64
-		{013E1B04-E320-4AB6-97DA-D52238E83783}.Debug|x64.Build.0 = Debug|x64
-		{013E1B04-E320-4AB6-97DA-D52238E83783}.Debug|x86.ActiveCfg = Debug|Win32
-		{013E1B04-E320-4AB6-97DA-D52238E83783}.Debug|x86.Build.0 = Debug|Win32
-		{013E1B04-E320-4AB6-97DA-D52238E83783}.Release|x64.ActiveCfg = Release|x64
-		{013E1B04-E320-4AB6-97DA-D52238E83783}.Release|x64.Build.0 = Release|x64
-		{013E1B04-E320-4AB6-97DA-D52238E83783}.Release|x86.ActiveCfg = Release|Win32
-		{013E1B04-E320-4AB6-97DA-D52238E83783}.Release|x86.Build.0 = Release|Win32
+		{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}.Debug|x64.ActiveCfg = Debug|x64
+		{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}.Debug|x64.Build.0 = Debug|x64
+		{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}.Debug|x86.ActiveCfg = Debug|Win32
+		{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}.Debug|x86.Build.0 = Debug|Win32
+		{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}.Release|x64.ActiveCfg = Release|x64
+		{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}.Release|x64.Build.0 = Release|x64
+		{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}.Release|x86.ActiveCfg = Release|Win32
+		{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A57E7F80-6272-4275-8910-42243E696394}
+		SolutionGuid = {28C2733B-F068-4841-AF69-67D86117968E}
 	EndGlobalSection
 EndGlobal
