--- ClassBooking/ClassBooking.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.11.35312.102
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ClassBooking", "ClassBooking.vcxproj", "{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}.Debug|x64.ActiveCfg = Debug|x64
-		{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}.Debug|x64.Build.0 = Debug|x64
-		{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}.Debug|x86.ActiveCfg = Debug|Win32
-		{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}.Debug|x86.Build.0 = Debug|Win32
-		{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}.Release|x64.ActiveCfg = Release|x64
-		{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}.Release|x64.Build.0 = Release|x64
-		{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}.Release|x86.ActiveCfg = Release|Win32
-		{FD2DC51C-1D68-4C30-803A-31CA1775E9A2}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {28C2733B-F068-4841-AF69-67D86117968E}
-	EndGlobalSection
-EndGlobal
