--- ClassBooking/ClassBooking/x64/Debug/ClassBooking.exe.recipe
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project>
-  <ProjectOutputs>
-    <ProjectOutput>
-      <FullPath>C:\Users\jaewo\Desktop\25_1\전기프\ClassBooking\x64\Debug\ClassBooking.exe</FullPath>
-    </ProjectOutput>
-  </ProjectOutputs>
-  <ContentFiles />
-  <SatelliteDlls />
-  <NonRecipeFileRefs />
-</Project>
