--- ClassBooking/ClassBooking/x64/Debug/ClassBooking.tlog/ClassBooking.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.41.34120:TargetPlatformVersion=10.0.22621.0:
-Debug|x64|C:\Users\jaewo\Desktop\25_1\전기프\ClassBooking\|
