--- ClassBooking/classroom.txt
@@ -1,5 +1,0 @@
-301    1    09:00    17:00
-302    2    10:00    12:00      # 예약 가능 여부 오류
-303    1    18:00    16:00      # 시작 > 종료
-304    1    09:AA    10:00      # 잘못된 시간 형식
-aaa    1    10:00    11:00      # 잘못된 강의실 번호
