--- ClassBooking/correctCase/classroom.txt
@@ -1,3 +1,0 @@
-301    1    09:00    17:00
-302    1    10:00    18:00
-303    0    09:00    12:00
