--- ClassBooking/reservation.txt
@@ -1,6 +1,0 @@
-john123    301    08:00    09:00      # 강의실 시간 범위 초과
-guest      302    10:00    12:00      # user.txt에 비정상 사용자
-ghost      303    11:00    13:00      # 존재하지 않는 사용자
-staff      301    10:00    12:00
-staff      301    11:00    13:00      # 중복 예약 시간
-staff      301    13:00    18:00      # 3시간 초과 예약
