+++ ClassBooking/x64/Debug/ClassBooking.log
@@ -1,0 +1,2 @@
+﻿  ClassBooking.cpp
+  ClassBooking.vcxproj -> C:\Users\jaewo\Desktop\25_1\전기프\ClassBooking\x64\Debug\ClassBooking.exe
