+++ ClassBooking/x64/Debug/ClassBooking.tlog/Cl.items.tlog
@@ -1,0 +1,1 @@
+C:\Users\jaewo\Desktop\25_1\전기프\ClassBooking\ClassBooking.cpp;C:\Users\jaewo\Desktop\25_1\전기프\ClassBooking\ClassBooking\x64\Debug\ClassBooking.obj
