+++ ClassBooking/x64/Debug/ClassBooking.tlog/link.secondary.1.tlog
@@ -1,0 +1,2 @@
+^C:\USERS\JAEWO\DESKTOP\25_1\전기프\CLASSBOOKING\CLASSBOOKING\X64\DEBUG\CLASSBOOKING.OBJ
+C:\Users\jaewo\Desktop\25_1\전기프\ClassBooking\ClassBooking\x64\Debug\ClassBooking.ilk
