+++ correctCase/reservation.txt
@@ -1,0 +1,3 @@
+john123    301    09:00    11:00
+guest1     302    10:00    11:00
+john123    301    13:00    14:00
