+++ correctCase/user.txt
@@ -1,0 +1,3 @@
+john123    Pass1234    0
+admin1     Admin@123   1
+guest1     hello123    0
